modifications

+++ approximative_solutions/src/solution/cplex.cc
@@ -74,17 +74,18 @@ void Cplex::Run() {
   // R_IDJPT => a tarefa I que esta na maquina J, comeca a ler o seu D-esimo dado de entrada
   // (note que nao eh o dado de indice D) a partir da maquina P no periodo T
   for (int i = 0; i < _n; i++) {
-    Task*              task        = GetTaskPerId(i + 1);
+    Task*              task        = GetTaskPerId(static_cast<size_t>(i + 1));
     std::vector<File*> input_files = task->get_input_files();
-    cplx.r[i]                      =  IloArray<IloArray<IloArray<IloBoolVarArray>>>(cplx.env, input_files.size());
+    cplx.r[i]                      =  IloArray<IloArray<IloArray<IloBoolVarArray>>>(cplx.env,
+        static_cast<int>(input_files.size()));
 
-    for(int j = 0; j < input_files.size(); j++) {
+    for (int j = 0; j < static_cast<int>(input_files.size()); j++) {
       cplx.r[i][j] = IloArray<IloArray<IloBoolVarArray>>(cplx.env, _m);
-      for(int k=0; k < _m; k++ ){
+      for (int k = 0; k < _m; k++) {
         cplx.r[i][j][k] = IloArray<IloBoolVarArray>(cplx.env, _m);
-        for(int l=0; l < _mb; l++) {
+        for (int l = 0; l < _mb; l++) {
           cplx.r[i][j][k][l] = IloBoolVarArray(cplx.env, _t);
-          for(int m=0; m < _t; m++) {
+          for (int m = 0; m < _t; m++) {
             sprintf (var_name, "r_%d_%d_%d_%d_%d", (int)i,(int)j, (int)k, (int)l, (int)m);            // nome da variavel
             cplx.r[i][j][k][l][m] = IloBoolVar(cplx.env, var_name);                                   // aloca variavel
             cplx.model.add(cplx.r[i][j][k][l][m]);                                                    // adiciona variavel ao modelo
@@ -93,7 +94,7 @@ void Cplex::Run() {
       }
     }
   }
-  
+
 
   IloCplex solver(cplx.model);                        // declara variável "solver" sobre o modelo a ser solucionado
   solver.exportModel("model.lp");                     // escreve modelo no arquivo no formato .lp
